tests/requirements.txt
pytest-cov
pytest-django
django_coverage_plugin==1.6.0
-e .  # installs wagtailaudioembed
